refactor(web): share metric-card primitives between detail panels
